Move cut finding into cutting package
